opencv: make opencv build with local ci configuration
On the current meta-openembedded master (changeset 9cd117a5dc),
opencv is blacklisted, as it does not compile with GCC6. As our
ci configuration uses GCC 5.3, we can un-blacklist opencv.

Also, opencv fails with some optional dependencies with the
PACKAGECONFIG set upstream. We do not require the PACKAGECONFIG
that cause failures and hence override the PACKAGECONFIG with the
one's we need and do not fail.

Signed-off-by: Lukas Bulwahn <lukas.bulwahn@oss.bmw-carit.de>