geographiclib: initial commit
The initial commit from Johannes Schrimpf provided the geographiclib
recipe in recipes-ros. The general policy is to only have ROS
packages in recipes-ros, i.e., packages that are tracked and updated
in the rosdistro/distribution.yaml file.
Hence, I then moved the geographiclib recipe to recipes-extended,
as it is a supporting library, not a ROS package.

Signed-off-by: Lukas Bulwahn <lukas.bulwahn@gmail.com>